pyproject.toml
[project]
name = "ssl-rfdetr-pneumonia"
version = "1.0.0"
description = "Domain-Adaptive SSL Continual Pre-training on DINOv2 Backbone for RF-DETR Pneumonia Detection"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "rfdetr",
    "albumentations>=1.4.0",
    "opencv-python-headless>=4.8.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "Pillow>=10.0.0",
    "wandb>=0.16.0",
    "pycocotools>=2.0.7",
    "scikit-learn>=1.3.0",
    "tqdm>=4.66.0",
    "pyyaml>=6.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "ruff>=0.1.0",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
